requirements.txt
flask
langchain
langchain-community
huggingface_hub
python-dotenv
